+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/spall/compare/v0.4.0...HEAD)
 ------------------------------------------------------------------------
 ### Added
+- Adds additional keyword arguments to control stdout and stderr
 - Includes `CalledProcessError` in `spall.exceptions`
 
 [0.4.0](https://github.com/jshwi/spall/releases/tag/v0.4.0) - 2022-05-09